fix: update release to handle prerelease
prev the release process was running some uv commands like `uv sync` and `uv build` without the `--prerelease allow` flag. That would change uv.lock as well as make a release bundle that did not depend on the expected things. This then caused subsequent builds to break (uv.lock was not valid anymore).

let's see if this fixes the release!