chore: schedule combine-pr github action
The goal is to automatically create PRs like this one: https://github.com/aptible/unpage/pull/132

The PR will still need to be reviewed by one of the aptible/unpage maintainers, and one of the maintainers will have to manually take action so it passes the PR checks.

I set the job to run Dependabot on all the weekdays first thing in the morning (for North America timezones), because dependabot doesn't seem to run on a consistent day each week.